seed games and protected endpoints

+++ transcendence-app/src/game/game.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   HttpStatus,
@@ -25,23 +26,43 @@ import {
   ApiTags,
   ApiUnprocessableEntityResponse,
 } from '@nestjs/swagger';
+import { User as GetUser } from '../user/decorators/user.decorator';
 import { Game } from './infrastructure/db/game.entity';
 import { MAX_ENTRIES_PER_PAGE } from '../shared/constants';
 import { PaginationWithSearchQueryDto } from '../shared/dtos/pagination-with-search.query.dto';
 import { CreateGameDto } from './dto/create-game.dto';
+import { CaslAbilityFactory } from '../authorization/casl-ability.factory';
+import { AuthorizationService } from '../authorization/authorization.service';
+import { User } from '../user/infrastructure/db/user.entity';
+import { Action } from '../shared/enums/action.enum';
 
 @Controller()
 @UseGuards(TwoFactorAuthenticatedGuard)
 @ApiTags('game')
 @ApiForbiddenResponse({ description: 'Forbidden' })
 export class GameController {
-  constructor(private gameService: GameService) {}
+  constructor(
+    private gameService: GameService,
+    private readonly authorizationService: AuthorizationService,
+    private readonly caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   @Post('game')
   @ApiCreatedResponse({ description: 'Add a game', type: Game })
   @ApiBadRequestResponse({ description: 'Bad Request' })
   @ApiUnprocessableEntityResponse({ description: 'Unprocessable entity' })
-  async addGame(@Body() game: CreateGameDto): Promise<Game> {
+  async addGame(
+    @GetUser() user: User,
+    @Body() game: CreateGameDto,
+  ): Promise<Game> {
+    const userWithAuthorization =
+      await this.authorizationService.getUserWithAuthorizationFromId(user.id);
+    const ability = this.caslAbilityFactory.defineAbilitiesFor(
+      userWithAuthorization,
+    );
+    if (ability.cannot(Action.Create, 'all')) {
+      throw new ForbiddenException();
+    }
     const savedGame = await this.gameService.addGame(game);
 
     if (!savedGame) {
@@ -56,7 +77,15 @@ export class GameController {
   @ApiBadRequestResponse({ description: 'Bad Request' })
   @ApiNotFoundResponse({ description: 'Not Found' })
   @ApiServiceUnavailableResponse({ description: 'Service Unavailable' })
-  async removeGame(gameId: string): Promise<void> {
+  async removeGame(@GetUser() user: User, gameId: string): Promise<void> {
+    const userWithAuthorization =
+      await this.authorizationService.getUserWithAuthorizationFromId(user.id);
+    const ability = this.caslAbilityFactory.defineAbilitiesFor(
+      userWithAuthorization,
+    );
+    if (ability.cannot(Action.Delete, 'all')) {
+      throw new ForbiddenException();
+    }
     return this.gameService.deleteGame(gameId);
   }
